requirements.txt
Django==2.2.4
django-pandas==0.6.1
djangorestframework==3.10.2
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.17.0
openpyxl==2.6.2
pandas==0.25.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
